requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
qrcode[pil]==7.4.2
Pillow==10.1.0
python-dotenv==1.0.0
